+++ C++/SRP/src/sha/BinaryString.h
@@ -26,7 +26,7 @@ class BinaryString {
 		BinaryWord* get_32b_word_at(int chunk_nr, int word_nr);
 
 		void get_words_from_chunk(unsigned int chunk_nr, std::vector<BinaryWord*>& out);
-		void get_words_from_chunk(unsigned int chunk_nr, std::vector<std::uint64_t>& out);
+		void get_words_from_chunk(unsigned int chunk_nr, std::vector<std::uint32_t>& out);
 
 		virtual unsigned int get_num_512b_chunks() const;
 
@@ -34,7 +34,7 @@ class BinaryString {
 		BinaryString();
 
 		std::vector<BinaryWord*> words;
-		std::vector<std::uint64_t> uint_words;
+		std::vector<std::uint32_t> uint_words;
 		std::string message;
 		unsigned int num_512b_chunks;
 };
